Travis: Raising exception in testharness
If a warning or failure is encountered in any of the tests,
an exception is raised and the process exists with the exit
code 1, else it exists with exit code 0.
Exiting with exit code 1 is crucial since it is picked up by
travis. This way travis will mark "make test" as failed.

This closes #39.