TestHarness: Spawning new process for each test.
Before running a test a new process is spawned in testharness that writes to stdout after a set amount of time (here 7min) to let Travis know the process is still running. After completion of a test, the spawned process is terminated. This procedure is repeated for each test.
This leads to a slightly more elegant way of writing to stdout rather than writing in a set interval while running all tests.